# docs/openapi/pdt-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Platform Data Toolkit (PDT) API",
        "description": "PDT is a centralized API server that serves as an enterprise knowledge silo for company-specific concepts, documents, and their relationships.\n\n## Features\n\n- **Knowledge Asset Management**: Store and manage knowledge assets with rich metadata\n- **Tag-Based Classification**: Multi-dimensional tagging including asset type, business domain, language, etc.\n- **Graph-Based Relationships**: Define and traverse relationships between assets\n- **Concept Collections**: Organize assets into named collections\n- **Search & Discovery**: Full-text search and tag-based filtering\n- **Audit Logging**: Track all changes with user attribution\n- **Authorization Context**: Cedar-based visibility, owner groups, and confidentiality management",
        "contact": {
            "name": "PDT Team"
        },
        "license": {
            "name": "MIT OR Apache-2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/assets": {
            "get": {
                "tags": [
                    "assets"
                ],
                "summary": "List assets with pagination",
                "operationId": "list_assets",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "asset_type",
                        "in": "query",
                        "description": "Filter by asset type tag value (e.g., \"document\", \"concept\", \"idea\")",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field: \"created_at\" or \"updated_at\" (default: \"updated_at\")",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order: \"asc\" or \"desc\" (default: \"desc\")",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of assets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse_Asset"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "assets"
                ],
                "summary": "Create a new asset",
                "operationId": "create_asset",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAssetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Asset created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Asset"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/assets/{id}": {
            "get": {
                "tags": [
                    "assets"
                ],
                "summary": "Get asset by ID",
                "operationId": "get_asset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Asset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "assets"
                ],
                "summary": "Update an asset",
                "operationId": "update_asset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAssetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Asset updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Asset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "assets"
                ],
                "summary": "Delete an asset (soft delete)",
                "operationId": "delete_asset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset deleted successfully"
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            }
        },
        "/api/assets/{id}/auth-context": {
            "put": {
                "tags": [
                    "assets"
                ],
                "summary": "Update the authorization context of an asset (Cedar visibility/ownership management)",
                "operationId": "update_auth_context",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAuthContextRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Auth context updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Asset"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            }
        },
...
        "/api/assets/{id}/tags": {
            "post": {
                "tags": [
                    "assets"
                ],
                "summary": "Add a tag to an asset",
                "operationId": "add_tag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddTagRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Tag added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tag"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            }
        },
        "/api/assets/{id}/tags/{tag_id}": {
            "delete": {
                "tags": [
                    "assets"
                ],
                "summary": "Remove a tag from an asset",
                "operationId": "remove_tag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag_id",
                        "in": "path",
                        "description": "Tag ID to remove",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag removed successfully"
                    },
                    "404": {
                        "description": "Asset or tag not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AddTagRequest": {
                "type": "object",
                "description": "Request to add a tag to an asset",
                "required": [
                    "category",
                    "value"
                ],
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
            "Asset": {
                "type": "object",
                "description": "Knowledge asset document",
                "required": [
                    "_id",
                    "title",
                    "created_at",
                    "updated_at",
                    "created_by",
                    "updated_by"
                ],
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "auth_context": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/AuthContext",
                                "description": "Cedar authorization context \u2014 when present, Cedar policies are enforced.\nWhen absent (grandfathered assets), access is open."
                            }
                        ]
                    },
                    "content": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "metadata": {
                        "type": "object",
                        "additionalProperties": {},
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    },
                    "title": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_by": {
                        "type": "string"
                    }
                }
            },
...
            "AuthContext": {
                "type": "object",
                "description": "Authorization context embedded on an asset.\n\nThree fields that don't exist anywhere in Asset or Tags.\nThis struct is 1:1 with the Asset document (MongoDB-idiomatic).",
                "properties": {
                    "confidentiality": {
                        "type": "string",
                        "description": "Security classification level: \"internal\", \"confidential\", \"restricted\""
                    },
                    "owner_groups": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Groups that have implicit access beyond the visibility rules"
                    },
                    "visibility": {
                        "type": "string",
                        "description": "Authorization visibility level: \"public\", \"team\", \"private\", \"org\""
                    }
                }
            },
...
            "CreateAssetRequest": {
                "type": "object",
                "description": "Request to create a new asset\nNote: Asset type should be specified via tags with category \"asset_type\"",
                "required": [
                    "title"
                ],
                "properties": {
                    "auth_context": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/AuthContext",
                                "description": "Optional auth_context to set at creation time (overrides defaults).\nIf omitted, defaults to public/internal/empty."
                            }
                        ]
                    },
                    "content": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "metadata": {
                        "type": "object",
                        "additionalProperties": {},
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AddTagRequest"
                        }
                    },
                    "title": {
                        "type": "string"
                    }
                }
            },
...
            "PaginatedResponse_Asset": {
                "type": "object",
                "description": "Paginated response wrapper",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "description": "Knowledge asset document",
                            "required": [
                                "_id",
                                "title",
                                "created_at",
                                "updated_at",
                                "created_by",
                                "updated_by"
                            ],
                            "properties": {
                                "_id": {
                                    "type": "string"
                                },
                                "auth_context": {
                                    "oneOf": [
                                        {
                                            "type": "null"
                                        },
                                        {
                                            "$ref": "#/components/schemas/AuthContext",
                                            "description": "Cedar authorization context \u2014 when present, Cedar policies are enforced.\nWhen absent (grandfathered assets), access is open."
                                        }
                                    ]
                                },
                                "content": {
                                    "type": [
                                        "string",
                                        "null"
                                    ]
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "created_by": {
                                    "type": "string"
                                },
                                "deleted_at": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "format": "date-time"
                                },
                                "metadata": {
                                    "type": "object",
                                    "additionalProperties": {},
                                    "propertyNames": {
                                        "type": "string"
                                    }
                                },
                                "tags": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tag"
                                    }
                                },
                                "title": {
                                    "type": "string"
                                },
                                "updated_at": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "updated_by": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "next_cursor": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "total": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64",
                        "minimum": 0
                    }
                }
            },
...
            "Tag": {
                "type": "object",
                "description": "Tag attached to an asset\n\nUses industry-standard key-value format:\n- `category`: Free-form string (e.g., \"type\", \"status\", \"domain\", \"priority\")\n- `value`: The tag value (e.g., \"document\", \"draft\", \"backend\", \"high\")\n\nCommon categories (suggestions, not enforced):\n- `type` - Asset type (document, idea, requirement, bug, task)\n- `status` - Status (draft, review, approved, archived)\n- `domain` - Business domain (backend, frontend, infra)\n- `lang` - Language (en, fa)\n- `priority` - Priority (high, medium, low)\n- `scope` - Feature/component scope",
                "required": [
                    "id",
                    "category",
                    "value",
                    "added_by",
                    "added_at"
                ],
                "properties": {
                    "added_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "added_by": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
...
            "UpdateAssetRequest": {
                "type": "object",
                "description": "Request to update an existing asset",
                "properties": {
                    "content": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "metadata": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "additionalProperties": {},
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "title": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
            "UpdateAuthContextRequest": {
                "type": "object",
                "description": "Request to update the authorization context of an asset",
                "properties": {
                    "confidentiality": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "owner_groups": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {
                            "type": "string"
                        }
                    },
                    "visibility": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
...
        }
    },
...
}